+++ src/crypto/abe/keygen/cover_crypt/cover_crypt_keygen.ts
@@ -1,5 +1,5 @@
 /* tslint:disable:max-classes-per-file */
-import { webassembly_generate_master_keys, webassembly_generate_user_private_key } from "../../../../../wasm_lib/abe/cover_crypt"
+import { webassembly_generate_master_keys, webassembly_generate_user_private_key, webassembly_rotate_attributes } from "../../../../../wasm_lib/abe/cover_crypt"
 import { logger } from "../../../../utils/logger"
 import { fromBeBytes } from "../../../../utils/utils"
 import { AbeKeyGeneration, AbeMasterKey } from "../keygen"
@@ -32,6 +32,17 @@ export class CoverCryptMasterKeyGeneration extends AbeKeyGeneration {
     const userPrivateKey = webassembly_generate_user_private_key(privateKey, accessPolicy, policyBytes)
 
     return userPrivateKey
+  }
+
+  public rotateAttributes(attributes: string[], policy: Policy): Policy {
+    logger.log(() => "attributes: " + attributes)
+    logger.log(() => "policy: " + policy)
+
+    const policyBytes = policy.toJsonEncoded()
+    const attributesBytes = new TextEncoder().encode(JSON.stringify(attributes))
+    const newPolicyString = webassembly_rotate_attributes(attributesBytes, policyBytes)
+
+    return Policy.fromJsonEncoded(newPolicyString)
   }
 
 }