+++ src/crypto/abe/keygen/gpsw/gpsw_crypt_keygen.ts
@@ -1,7 +1,11 @@
 /* tslint:disable:max-classes-per-file */
-import { webassembly_generate_master_keys, webassembly_generate_user_private_key } from "../../../../../wasm_lib/abe/gpsw/abe_gpsw"
+import {
+  webassembly_generate_master_keys,
+  webassembly_generate_user_private_key,
+  webassembly_rotate_attributes
+} from "../../../../../wasm_lib/abe/gpsw/abe_gpsw"
 import { logger } from "../../../../utils/logger"
-import { fromBeBytes } from "../../../../utils/utils"
+import { fromBeBytes, hexDecode } from "../../../../utils/utils"
 import { AbeKeyGeneration, AbeMasterKey } from "../keygen"
 import { Policy } from "../policy"
 
@@ -32,6 +36,18 @@ export class GpswMasterKeyGeneration extends AbeKeyGeneration {
     const userPrivateKey = webassembly_generate_user_private_key(privateKey, accessPolicy, policyBytes)
 
     return userPrivateKey
+  }
+
+  public rotateAttributes(attributes: string[], policy: Policy): Policy {
+    logger.log(() => "attributes: " + attributes)
+    logger.log(() => "policy: " + policy)
+
+    const policyBytes = policy.toJsonEncoded()
+    const attributesBytes = new TextEncoder().encode(JSON.stringify(attributes))
+    const newPolicyBytes = webassembly_rotate_attributes(attributesBytes, policyBytes)
+
+    const newPolicyString = new TextDecoder().decode(hexDecode(newPolicyBytes))
+    return Policy.fromJsonEncoded(newPolicyString)
   }
 
 }