+++ src/crypto/abe/keygen/keygen.ts
@@ -31,5 +31,6 @@ export class AbeMasterKey {
 export abstract class AbeKeyGeneration {
   public abstract generateMasterKey(policy: Policy): AbeMasterKey
   public abstract generateUserPrivateKey(privateKey: Uint8Array, accessPolicy: string, policy: Policy): Uint8Array
+  public abstract rotateAttributes(attributes: string[], policy: Policy): Policy
 
 }