+++ CHANGELOG.md
@@ -6,8 +6,9 @@ All notable changes to this project will be documented in this file.
 ## [1.0.1] - 2022-06-08
 ### Added
 - Add ABE key generation: master key generation and user decryption key generation
+- Add attributes rotation
 ### Changed
-- Depends on `abe_gpsw` v0.6.11 and `cover_crypt` v2.0.1
+- Depends on `abe_gpsw` v0.8.0 and `cover_crypt` v3.0.1
 ### Fixed
 ### Removed